feat: [PIDM-1516] - add new transaction controller for API and update openapi
